pyproject.toml
[project]
name = "portfolio-data-quality-pydantic-and-tdd"
version = "0.1.0"
description = ""
authors = [
    {name = "Humberto Lomeu",email = "hcslomeu@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "sqlalchemy (>=2.0.45,<3.0.0)",
    "pydantic[email,timezone] (>=2.12.5,<3.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "fastapi (>=0.124.2,<0.125.0)",
    "psycopg (>=3.3.2,<4.0.0)",
    "mkdocs (>=1.6.1,<2.0.0)",
    "pandas (>=2.3.3,<3.0.0)",
    "streamlit (>=1.52.1,<2.0.0)",
    "mkdocs-mermaid2-plugin (>=1.2.3,<2.0.0)",
    "mkdocs-material (>=9.7.0,<10.0.0)",
    "mkdocstrings-python (>=2.0.1,<3.0.0)",
    "taskipy (>=1.14.1,<2.0.0)",
    "isort (>=7.0.0,<8.0.0)",
    "black (>=25.12.0,<26.0.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pandera[io] (>=0.27.0,<0.28.0)",
    "duckdb (>=1.4.3,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.taskipy.tasks]
format = """
isort .
black .
"""
kill = "kill -9 $(lsof -t -i :8000)"
test = "pytest -v"
run = """
python app/main.py
"""
doc = "mkdocs serve"
